'soitemid-datatype-changed'

+++ AddChipsPackingForm.cs
@@ -42,7 +42,7 @@ namespace PackingApplication
         LotsResponse lotResponse = new LotsResponse();
         WeighingScaleReader wtReader = new WeighingScaleReader();
         string comPort;
-        int selectedSOId = 0;
+        long selectedSOId = 0;
         decimal totalSOQty = 0;
         decimal totalProdQty = 0;
         int selectLotId = 0;
